refactor: 物件詳細ページ読みやすいようにコード修正
