+++ frontend/components/features/property/PropertyDetail.tsx
@@ -1,12 +1,8 @@
 "use client";
-import React, { ReactNode, useEffect, useState } from "react";
+import React from "react";
 import { AmenitiesProps, PropertyDetailData } from "@/types/notionTypes";
 import {
-  ArrowRight,
-  Check,
-  Copy,
   Dumbbell,
-  Instagram,
   KeyRound,
   Plug,
   Thermometer,
@@ -15,29 +11,19 @@ import {
   Wifi,
 } from "lucide-react";
 import {
-  Breadcrumb,
-  BreadcrumbList,
-  BreadcrumbItem,
-  BreadcrumbSeparator,
-  BreadcrumbPage,
-  BreadcrumbLink,
-} from "@/components/ui/breadcrumb";
-import Link from "next/link";
-import {
   AccessMap,
   Amenities,
   BasicInfo,
   Conditions,
+  ContactCard,
+  ContactPopUpPC,
+  ContactPopUpSP,
   InstagramAds,
   Neighbors,
   PropertyImage,
-  SectionWrapper,
-  StaffComment,
+  RecommendPoint,
+  BreadcrumbArea,
 } from "@/components/features/property/PropertyDetailContents";
-import Image from "next/image";
-import { Input } from "@/components/ui/input";
-import { toast } from "sonner";
-import { useRouter } from "next/navigation";
 import { formatMoveInDate } from "@/utlis/getPropertyValue";
 import {
   DetailPageHeader,
@@ -51,7 +37,7 @@ const propertyDetailHeaderList: DetailPageHeaderOption[] = [
   { id: "basic-info", title: "基本情報" },
   { id: "conditions", title: "入居条件" },
   { id: "facilities", title: "設備" },
-  { id: "comment", title: "おすすめポイント" },
+  { id: "point", title: "おすすめポイント" },
   { id: "map", title: "アクセスマップ" },
   { id: "neighbors", title: "周辺情報" },
   { id: "contact", title: "お問い合わせ" },
@@ -142,133 +128,33 @@ const PropertyDetail = ({ property }: { property: PropertyDetailData }) => {
     },
   ];
 
-  // ヘッダーから遷移できる情報エリアの表示用
-  const propertyInfoList: { id: string; title: string; body: ReactNode }[] = [
-    {
-      id: "images",
-      title: "",
-      body: (
-        <PropertyImage
-          imgUrl={property.thumbnail}
-          title={property.title}
-          googlePhotoUrl={property.image}
-        />
-      ),
-    },
-    {
-      id: "basic-info",
-      title: "基本情報",
-      body: (
-        <BasicInfo
-          rent={property.rent}
-          deposit={property.deposit}
-          area={property.area}
-          closestStation={property.closestStation}
-          zone={property.zone}
-          moveInDate={moveInDate}
-        />
-      ),
-    },
-    {
-      id: "conditions",
-      title: "入居条件",
-      body: <Conditions conditions={conditions} />,
-    },
-    {
-      id: "facilities",
-      title: "設備",
-      body: <Amenities amenities={amenities} />,
-    },
-    {
-      id: "comment",
-      title: "おすすめポイント",
-      body: <StaffComment comment={property.staffComment} />,
-    },
-    {
-      id: "map",
-      title: "アクセスマップ",
-      body: (
-        <AccessMap
-          geoPosition={property.geoPosition}
-          closestStation={property.closestStation}
-          timeToStation={property.timeToStation}
-        />
-      ),
-    },
-    { id: "neighbors", title: "周辺情報", body: <Neighbors /> },
-    // {
-    //   id: "available-properties",
-    //   title: "近隣の空き物件",
-    //   body: <AvailableProperties />,
-    // },
-  ];
-
-  const router = useRouter();
-
-  /**
-   * パンクズリスト物件一覧に戻る
-   * @param e {React.MouseEvent<HTMLAnchorElement>}
-   */
-  const handleBack = (e: React.MouseEvent<HTMLAnchorElement>) => {
-    e.preventDefault(); // リンクのデフォルト動作を無効化
-
-    // 履歴がある場合は戻る
-    if (window.history.length > 2) {
-      router.back();
-    } else {
-      // 外部ページから遷移した場合等、履歴がない場合は特定のURLに遷移
-      router.push("/properties");
-    }
-  };
-
   return (
     <DetailPageFrame>
       <DetailPageHeader headerOptions={propertyDetailHeaderList} />
       {/* --- 左上エリア：　パンクズリスト　--- 　*/}
       <div className="pt-5 text-sm">
-        <Breadcrumb>
-          <BreadcrumbList>
-            {/* ルートにホームページを作ったらこれ使うのでそれまで非表示 */}
-            {/* <BreadcrumbItem>
-              <BreadcrumbLink>
-                <Link href="/">Home</Link>
-              </BreadcrumbLink>
-            </BreadcrumbItem>
-            <BreadcrumbSeparator /> */}
-            <BreadcrumbItem>
-              <BreadcrumbLink onClick={handleBack} className="cursor-pointer">
-                物件一覧
-              </BreadcrumbLink>
-            </BreadcrumbItem>
-            <BreadcrumbSeparator />
-            <BreadcrumbItem>
-              <BreadcrumbPage className="text-gray-500">
-                {property.roomName}
-              </BreadcrumbPage>
-            </BreadcrumbItem>
-          </BreadcrumbList>
-        </Breadcrumb>
+        <BreadcrumbArea label={property.roomName} />
       </div>
 
       <div className="flex flex-col sm:flex-row sm:justify-between items-center gap-3 py-3 sm:py-6">
-        {/* ---　　タイトルエリア --- */}
         <div className="flex flex-col items-start gap-2 w-full sm:w-[50%] lg:w-[60%]">
           {/* 空き状況ステータス */}
           <div className="py-1.5 px-4 bg-white rounded-full z-40 top-3 left-3 shadow-lg opacity-85 text-xs sm:text-sm border border-themeColor text-themeColor">
             {property.status}
           </div>
-
+          {/* タイトル */}
           <h1 className="text-base sm:text-base lg:text-2xl font-semibold">
             {property.title}
           </h1>
         </div>
 
-        {/* ---　　右上エリア：　家賃と入居日 --- */}
         <div className="flex gap-7 tracking-widest font-semibold">
+          {/* 家賃 */}
           <div className="flex flex-col items-center">
             <div className="text-gray-400 text-sm">MONTHLY RENT</div>
             <div className="text-sm sm:text-base">CA${property.rent}</div>
           </div>
+          {/* 入居可能日 */}
           <div className="flex flex-col items-center">
             <div className="text-gray-400 text-sm">MOVE IN</div>
             <div className="text-sm sm:text-base">{moveInDate}</div>
@@ -279,32 +165,47 @@ const PropertyDetail = ({ property }: { property: PropertyDetailData }) => {
       {/* --- ヘッダーから遷移できる情報エリア --- */}
       <div className="lg:h-[60lvh] flex flex-col lg:flex-row lg:justify-between mx-auto">
         <div className="w-full lg:w-[60%] lg:pr-5">
-          {propertyInfoList.map((info) => (
-            <div id={info.id} key={info.id} className="border-b">
-              <>
-                {info.title && <SectionTitle title={info.title} />}
-                <>{info.body}</>
-              </>
-            </div>
-          ))}
+          {/* 画像 */}
+          <PropertyImage
+            imgUrl={property.thumbnail}
+            title={property.title}
+            googlePhotoUrl={property.image}
+          />
+          {/* 基本情報 */}
+          <BasicInfo
+            rent={property.rent}
+            deposit={property.deposit}
+            area={property.area}
+            closestStation={property.closestStation}
+            zone={property.zone}
+            moveInDate={moveInDate}
+          />
+          {/* 入居条件 */}
+          <Conditions conditions={conditions} />
+          {/* 設備 */}
+          <Amenities amenities={amenities} />
+          {/* おすすめポイント */}
+          <RecommendPoint text={property.staffComment} />
+          {/* アクセスマップ */}
+          <AccessMap
+            geoPosition={property.geoPosition}
+            closestStation={property.closestStation}
+            timeToStation={property.timeToStation}
+          />
+          {/* 周辺情報 */}
+          <Neighbors />
         </div>
 
         <div className="w-full lg:w-[40%] xl:pl-14">
-          {/* お問い合わせエリア */}
-          <div id="contact" className="border-b lg:border-none">
-            <div className="text-base sm:text-xl font-semibold tracking-wides pt-8 lg:pt-0">
-              お問い合わせ
-            </div>
-            <ContactCard />
-          </div>
-          {/* サービス内容紹介エリア */}
-          <div id="about-us">
-            <SectionTitle title="バンクーバーのお家について" />
-            <InstagramAds />
-          </div>
+          {/* お問い合わせ */}
+          <ContactCard />
+
+          {/* バンクーバーのお家について*/}
+          <InstagramAds />
         </div>
       </div>
 
+      {/*　画面右下まずは無料相談 */}
       <div className="hidden sm:block">
         <ContactPopUpPC />
       </div>
@@ -316,145 +217,3 @@ const PropertyDetail = ({ property }: { property: PropertyDetailData }) => {
 };
 
 export default PropertyDetail;
-
-/**
- * 項目のタイトルコンポーネント
- *
- * @param title {string}
- */
-const SectionTitle = ({ title }: { title: string }) => {
-  return (
-    <h5 className="text-base sm:text-xl font-semibold tracking-widest pt-8">
-      {title}
-    </h5>
-  );
-};
-
-/**
- * 右下、『まずは無料相談』コンポーネント（パソコン）
- *
- */
-const ContactPopUpPC = () => {
-  return (
-    <Link
-      className="fixed sm:bottom-10 sm:right-10 h-32 w-32 rounded-full bg-red-600 text-white flex flex-col items-center justify-center gap-1 tracking-widest hover:scale-105 z-50"
-      href="https://www.instagram.com/vancouver.no.ouchi/"
-      target="_blank"
-    >
-      <div>まずは</div>
-      <div className="font-semibold">無料相談</div>
-      <ArrowRight />
-    </Link>
-  );
-};
-
-/**
- * 右下、『まずは無料相談』コンポーネント（スマホ）
- *
- */
-const ContactPopUpSP = () => {
-  return (
-    <Link
-      className="fixed bottom-0 right-0 bg-red-600 text-white flex items-center gap-1 tracking-widest p-3 rounded-tl-lg z-50"
-      href="https://www.instagram.com/vancouver.no.ouchi/"
-      target="_blank"
-    >
-      <div className="font-semibold">まずは無料相談</div>
-      <ArrowRight />
-    </Link>
-  );
-};
-
-/**
- * お問い合わせコンポーネント
- *
- */
-const ContactCard = () => {
-  const [url, setUrl] = useState("");
-  const [isCopied, setIsCopied] = useState<boolean>(false);
-
-  // クライアントサイドでwindowにアクセスする方法
-  useEffect(() => {
-    setUrl(window.location.href);
-  }, []);
-
-  const copyUrl = async () => {
-    try {
-      await navigator.clipboard.writeText(url);
-      setIsCopied(true);
-    } catch (error) {
-      console.error(error);
-      toast.error("URLのコピーに失敗しました。", { position: "top-right" });
-    }
-  };
-
-  useEffect(() => {
-    if (isCopied) {
-      const timer = setTimeout(() => {
-        setIsCopied(false);
-      }, 3000);
-
-      return () => clearTimeout(timer);
-    }
-  }, [isCopied]);
-
-  return (
-    <SectionWrapper>
-      <div className="rounded-lg px-4 py-8 sm:px-8 sm:py-12 shadow-xl border-[1px] border-gray-300">
-        <div className="w-full flex justify-center items-center">
-          <Image
-            src="/vancouver_no_ouchi_logo2.png"
-            alt="バンクーバーのお家ロゴ"
-            className="pr-4"
-            loading="lazy"
-            unoptimized={true}
-            width={90}
-            height={80}
-            style={{ objectFit: "cover" }}
-          />
-          <div className="flex flex-col items-center justify-center gap-1">
-            <p className="text-sm">シェアハウス探しは</p>
-            <p className="font-bold">バンクーバーのお家</p>
-            <p className="text-sm">にお任せください</p>
-          </div>
-        </div>
-        <div className="text-sm py-2 text-center text-themeColor font-semibold">
-          \ お問い合わせはこちらから /
-        </div>
-        <div className="flex flex-col gap-4">
-          <div>
-            <div className="pb-2 text-sm">①この物件のリンクをコピー</div>
-            <div className="relative">
-              <Input
-                value={url}
-                onChange={() => {}} // これがないとWarningが出る
-                className="pr-14 text-gray-500 border border-themeColor"
-              />
-              <div className="absolute right-3 top-3 flex items-center gap-1 text-sm text-themeColor">
-                {isCopied ? (
-                  <Check size={20} />
-                ) : (
-                  <Copy
-                    size={20}
-                    className="cursor-pointer"
-                    onClick={copyUrl}
-                  />
-                )}
-              </div>
-            </div>
-          </div>
-
-          <div>
-            <div className="pb-2 text-sm">②リンクをインスタのDMに送信</div>
-            <Link href="https://www.instagram.com/vancouver.no.ouchi">
-              <div className="flex items-center justify-center gap-3 w-full  bg-grayThemeColor text-themeColor border border-themeColor py-3 rounded-lg cursor-pointer hover:opacity-70">
-                <Instagram />
-                <div className="">DMで相談・内見予約</div>
-              </div>
-            </Link>
-          </div>
-        </div>
-      </div>
-    </SectionWrapper>
-  );
-};
